requirements.txt
inputs>=0.5
pyserial>=3.5
